+++ src/main/kotlin/com/challenge/country/CountryRepository.kt
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface CountryRepository : CrudRepository<Country, Long> {
+interface CountryRepository : CrudRepository<Country, Int> {
     fun existsByNameIgnoreCase(name: String): Boolean
 }
