+++ src/main/kotlin/com/challenge/state/RegisterStateRequestDto.kt
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.NotBlank
 class RegisterStateRequestDto(
     @field:NotBlank(message = "Name is required")
     val name: String,
-    val countryId: Long
+    val countryId: Int
 ) {
     fun toEntity(country: Country) = State(name = name.trim(), country = country)
 }
